+++ backend/apps/api/dataset/tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
 from apps.dataset.models import DataPoint, Dataset
-from apps.dataset.services.csv_parser import parse_row_time
+from apps.dataset.services.ingestion.csv_parser import parse_row_time
 
 
 @pytest.fixture
