+++ backend/apps/dataset/services/tasks_runner.py
@@ -1,5 +1,5 @@
 from apps.dataset.models import Dataset
-from apps.dataset.services.csv_parser import parse_dataset_csv
+from apps.dataset.services.ingestion.csv_parser import parse_dataset_csv
 
 
 def parse_dataset_sync(dataset_id: int):
